functions/cleaner_function/requirements.txt
azure-storage-blob==12.25.0
azure-functions
azure-ai-formrecognizer>=3.3.0
azure-openai==1.0.0
python-dotenv==1.0.0
openai>=1.12.0